Add files via upload
This is a Dots and Boxes game between a player and AI. The rules are simple, wherein each player connects 2 dots to make a line and continues to do so till a box is made by joining 4 lines. The person who connects the last line to complete the box earns the point. The player with maximum points ie with most boxes wins. 
The minimax algorithm along with alpha-beta pruning has been implemented for the computer to identify which move it should make next.